refactor:  separate login validation functions
