+++ src/pages/login/validation-chekcs/email-validation/constants.ts
@@ -1,0 +1,3 @@
+const EMAIL_SEPARATORS_CHECK =
+  /^([a-z0-9_-]+\.)*[a-z0-9_-]+@[a-z0-9_-]+(\.[a-z0-9_-]+)*\.[a-z]{2,6}$/;
+export default EMAIL_SEPARATORS_CHECK;
