+++ src/pages/login/validation-chekcs/helpers/check-spaces.ts
@@ -1,0 +1,8 @@
+import { ChangeEvent } from 'react';
+
+export default function checkSpaces(e: ChangeEvent<HTMLInputElement>) {
+  return (
+    e.target.value.toString().slice(0, 1) === ' ' ||
+    e.target.value.toString().slice(-1) === ' '
+  );
+}
