+++ src/pages/login/validation-chekcs/password-validation/constants.ts
@@ -1,0 +1,5 @@
+export const PASSWORD_NUMBER_CHECK = /(?=.*[0-9])/;
+export const PASSWORD_SPECIAL_SYMBOL_CHECK = /(?=.*[!@#$%^&*])/;
+export const PASSWORD_UPPER_CASE_SYMBOL_CHECK = /(?=.*[A-Z])/;
+export const PASSWORD_LOWER_CASE_SYMBOL_CHECK = /(?=.*[a-z])/;
+export const PASSWORD_SYMOLS_NUMBER_CHECK = /[0-9a-zA-Z!@#$%^&*]{6,}/;
