+++ src/pages/login/validation-chekcs/password-validation/helpers/change-password-visibility.ts
@@ -1,0 +1,11 @@
+export default function changePasswordVisibility(
+  e: React.MouseEvent<HTMLDivElement>
+) {
+  const passwordVisibility = (
+    (e.target as HTMLElement).previousSibling as HTMLInputElement
+  ).type;
+
+  ((e.target as HTMLElement).previousSibling as HTMLInputElement).type =
+    passwordVisibility === 'text' ? 'password' : 'text';
+  (e.target as HTMLElement).classList.toggle('visible');
+}
